+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/MockClientBuilder.kt
@@ -59,7 +59,7 @@ internal class MockClientBuilder {
             1000,
             false,
             ChatLogger.Config(ChatLogLevel.NOTHING, null),
-            ChatNotificationHandler(context)
+            ChatNotificationHandler(context),
         )
 
         socket = FakeChatSocket()